Refactor NameRecordHeader to name_resolver crate
