+++ .github/workflows/ci-stone-ios.yml
@@ -40,7 +40,6 @@ jobs:
 
       - name: Setup Xcode
         run: |
-          sudo xcode-select -s /Applications/Xcode_16.1.app
           xcrun simctl list devices
 
       - name: Test iOS
