+++ crates/gem_solana/src/model.rs
@@ -1,13 +1,8 @@
-use crate::pubkey::Pubkey;
-use borsh::{BorshDeserialize, BorshSerialize};
 use num_bigint::BigInt;
 use primitives::{AssetId, Chain};
 use serde::{Deserialize, Serialize};
 use serde_serializers::deserialize_biguint_from_str;
-use std::{
-    collections::{HashMap, HashSet},
-    time::SystemTime,
-};
+use std::collections::{HashMap, HashSet};
 
 pub use num_bigint::BigUint;
 
@@ -399,46 +394,4 @@ pub struct InflationRate {
 pub struct EpochInfo {
     pub epoch: u64,
     pub slots_in_epoch: u64,
-}
-
-#[derive(Debug, Clone, BorshSerialize, BorshDeserialize)]
-pub struct NameRecordHeader {
-    pub discriminator: [u8; 8],
-    pub parent_name: Pubkey,
-    pub owner: Pubkey,
-    pub nclass: Pubkey,
-    pub expires_at: u64,
-    pub created_at: u64,
-    pub non_transferable: u8,
-    pub padding: [u8; 79],
-}
-
-impl NameRecordHeader {
-    pub const BYTE_SIZE: usize = 8 + 32 + 32 + 32 + 8 + 8 + 1 + 79;
-    pub const EXPECTED_DISCRIMINATOR: [u8; 8] = [68, 72, 88, 44, 15, 167, 103, 243];
-
-    pub fn try_from_slice(data: &[u8]) -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
-        if data.len() < Self::BYTE_SIZE {
-            return Err("Invalid data length for NameRecordHeader".into());
-        }
-
-        let header: NameRecordHeader = BorshDeserialize::try_from_slice(data)?;
-
-        if header.discriminator != Self::EXPECTED_DISCRIMINATOR {
-            return Err("Invalid discriminator for NameRecordHeader".into());
-        }
-
-        Ok(header)
-    }
-
-    pub fn is_valid(&self) -> bool {
-        if self.expires_at == 0 {
-            return true;
-        }
-
-        let current_time = SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap().as_secs();
-        let grace_period = 45 * 24 * 60 * 60;
-
-        self.expires_at + grace_period > current_time
-    }
 }
