+++ crates/name_resolver/Cargo.toml
@@ -10,6 +10,7 @@ serde_json = { workspace = true }
 reqwest.workspace = true
 async-trait = { workspace = true }
 base64 = { workspace = true }
+borsh.workspace = true
 hex = { workspace = true }
 lazy_static = { workspace = true }
 alloy-primitives.workspace = true
