+++ crates/name_resolver/src/alldomains/mod.rs
@@ -1,0 +1,5 @@
+pub mod client;
+pub mod model;
+
+pub use client::AllDomainsClient;
+pub use model::NameRecordHeader;
